invisible-arr/edge-node/services/qa-swarm/requirements.txt
httpx>=0.28.0
asyncpg>=0.30.0
sqlalchemy[asyncio]>=2.0
playwright>=1.49.0
PyGithub>=2.5.0
prometheus-api-client>=0.5.0
python-json-logger>=3.2.0
